emergency-routing-model1/data/build_graph.py
"""Downloads Indian city road networks from OpenStreetMap via OSMnx and builds a graph structure for the GCN layer. Handles India-specific road features: speed breakers, unpaved roads, narrow lanes, and mixed road classes."""

import osmnx as ox
import networkx as nx
import numpy as np
import pandas as pd
import logging
from pathlib import Path
from scipy.sparse import save_npz, load_npz


def download_city_graph(city_name: str, network_type: str = "drive") -> nx.MultiDiGraph:
    """Download the road graph for a city from OpenStreetMap."""
    pass


def build_adjacency_matrix(graph: nx.MultiDiGraph) -> np.ndarray:
    """Build an adjacency matrix from a directed multigraph."""
    pass


def extract_node_features(graph: nx.MultiDiGraph) -> pd.DataFrame:
    """Extract node-level features required by the GCN layer."""
    pass


def assign_india_road_weights(graph: nx.MultiDiGraph) -> nx.MultiDiGraph:
    """Assign India-specific edge weights to the road graph."""
    pass


def save_graph(graph: nx.MultiDiGraph, output_dir: str) -> None:
    """Persist graph artifacts to disk."""
    pass


def load_graph(input_dir: str) -> nx.MultiDiGraph:
    """Load graph artifacts from disk."""
    pass
