emergency-routing-model1/data/fetch_incidents.py
"""Fetches real-time road incident data - accidents, closures, roadblocks - from Mappls Routing API and Waze CCP for Indian cities. Incident data is used as a feature spike signal in the LSTM model."""

import requests
import pandas as pd
import logging
from datetime import datetime
from dotenv import load_dotenv


def fetch_mappls_incidents(bbox: dict, api_key: str) -> pd.DataFrame:
    """Fetch incident data from the Mappls routing API."""
    pass


def fetch_waze_incidents(bbox: dict) -> pd.DataFrame:
    """Fetch incident data from the Waze CCP feed."""
    pass


def classify_incident_severity(incident_type: str) -> int:
    """Classify incident type into a discrete severity level."""
    pass


def compute_incident_impact_score(incidents_df: pd.DataFrame) -> pd.DataFrame:
    """Compute impact scores for incident records."""
    pass
