emergency-routing-model1/training/train.py
"""Training loop for the LSTM+GCN emergency traffic prediction model. Uses weighted MSE loss across prediction horizons (T+5, T+10, T+20, T+30 minutes). Includes early stopping, learning rate scheduling, and checkpoint saving."""

import torch
import torch.nn as nn
import torch.optim as optim
import numpy as np
import logging
import yaml
from pathlib import Path
from torch.utils.data import DataLoader, TensorDataset


def train_epoch(
    model: nn.Module,
    dataloader: DataLoader,
    optimizer: optim.Optimizer,
    criterion: nn.Module,
    device: torch.device,
) -> float:
    """Train the model for one epoch and return average loss."""
    pass


def validate_epoch(
    model: nn.Module,
    dataloader: DataLoader,
    criterion: nn.Module,
    device: torch.device,
) -> float:
    """Validate the model for one epoch and return average loss."""
    pass


def train(config: dict) -> None:
    """Run end-to-end training using configuration settings."""
    pass


def save_checkpoint(
    model: nn.Module,
    optimizer: optim.Optimizer,
    epoch: int,
    val_loss: float,
    path: str,
) -> None:
    """Save training checkpoint artifacts to disk."""
    pass


def load_checkpoint(
    model: nn.Module,
    optimizer: optim.Optimizer,
    path: str,
) -> tuple:
    """Load model and optimizer states from checkpoint."""
    pass


def compute_weighted_loss(
    predictions: torch.Tensor,
    targets: torch.Tensor,
    weights: list,
) -> torch.Tensor:
    """Compute horizon-weighted loss across prediction steps."""
    pass
