emergency-routing-model1/requirements.txt
# Core ML
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data
pandas>=2.0.0
pyarrow>=12.0.0
osmnx>=1.6.0
networkx>=3.1

# APIs and HTTP
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Training utilities
pyyaml>=6.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
httpx>=0.24.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
